+++ src/main/java/webserver/RequestHandler.java
@@ -11,7 +11,7 @@ import java.net.URISyntaxException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import model.HttpRequestFactory;
+import webserver.http.request.HttpRequestFactory;
 
 public class RequestHandler implements Runnable {
     private static final Logger logger = LoggerFactory.getLogger(RequestHandler.class);
