+++ src/main/java/webserver/http/FileType.java
@@ -1,4 +1,4 @@
-package model;
+package webserver.http;
 
 import java.util.Arrays;
