+++ src/main/java/webserver/http/request/HttpHeaderFields.java
@@ -1,4 +1,4 @@
-package model;
+package webserver.http.request;
 
 public final class HttpHeaderFields {
     public static final String CONTENT_LENGTH = "Content-Length";
