+++ src/main/java/webserver/http/request/HttpMethod.java
@@ -1,6 +1,10 @@
-package model;
+package webserver.http.request;
 
 import java.util.function.BiFunction;
+
+import webserver.http.service.AbstractHttpService;
+import webserver.http.service.HttpGetService;
+import webserver.http.service.HttpPostService;
 
 public enum HttpMethod {
     GET(HttpGetService::new),
