+++ src/main/java/webserver/http/request/HttpVersion.java
@@ -1,4 +1,4 @@
-package model;
+package webserver.http.request;
 
 import java.util.Arrays;
 import java.util.NoSuchElementException;
