+++ src/main/java/webserver/http/request/RequestURI.java
@@ -1,4 +1,6 @@
-package model;
+package webserver.http.request;
+
+import webserver.http.FileType;
 
 public class RequestURI {
     private final String uri;
