+++ src/main/java/webserver/http/request/RequestURIFactory.java
@@ -1,4 +1,4 @@
-package model;
+package webserver.http.request;
 
 @FunctionalInterface
 public interface RequestURIFactory {
