+++ src/main/java/webserver/http/service/AbstractHttpService.java
@@ -1,8 +1,11 @@
-package model;
+package webserver.http.service;
 
 import java.io.IOException;
 import java.io.OutputStream;
 import java.net.URISyntaxException;
+
+import webserver.http.request.HttpBody;
+import webserver.http.request.HttpHeader;
 
 public abstract class AbstractHttpService {
     protected final HttpHeader header;
