+++ src/main/java/webserver/http/service/HttpPostService.java
@@ -1,6 +1,4 @@
-package model;
-
-import static webserver.WebServer.*;
+package webserver.http.service;
 
 import java.io.DataOutputStream;
 import java.io.IOException;
@@ -10,6 +8,11 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import db.DataBase;
+import model.User;
+import webserver.http.request.HttpBody;
+import webserver.http.request.HttpHeader;
+import webserver.http.request.HttpParams;
+import webserver.http.request.HttpVersion;
 
 public class HttpPostService extends AbstractHttpService {
     private static final Logger logger = LoggerFactory.getLogger(HttpPostService.class);
@@ -30,7 +33,7 @@ public class HttpPostService extends AbstractHttpService {
         try {
             dos.writeBytes(HttpVersion.HTTP_1_1.getVersion() + " 302 Found \r\n");
             dos.writeBytes(
-                    "Location: http://" + BASE_URL + ":" + DEFAULT_PORT + "/index.html \r\n");
+                    "Location: http://localhost:8080/index.html \r\n");
             dos.writeBytes("\r\n");
             dos.flush();
         } catch (IOException e) {
